+++ common/src/main/java/net/frozenblock/wilderwild/block/PaleMushroomBlock.java
@@ -17,11 +17,11 @@
 
 package net.frozenblock.wilderwild.block;
 
+import net.frozenblock.lib.platform.ModLoader;
+import net.frozenblock.wilderwild.block.impl.client.PaleParticleSpawners;
 import net.frozenblock.wilderwild.registry.WWEnvironmentAttributes;
-import net.frozenblock.wilderwild.registry.WWParticleTypes;
 import net.minecraft.core.BlockPos;
 import net.minecraft.resources.ResourceKey;
-import net.minecraft.util.Mth;
 import net.minecraft.util.RandomSource;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.MushroomBlock;
@@ -29,11 +29,6 @@ import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.levelgen.feature.Feature;
 
 public class PaleMushroomBlock extends MushroomBlock {
-	public static final int MIN_PARTICLE_SPAWN_WIDTH = -3;
-	public static final int MAX_PARTICLE_SPAWN_WIDTH = 3;
-	public static final int MIN_PARTICLE_SPAWN_HEIGHT = -1;
-	public static final int MAX_PARTICLE_SPAWN_HEIGHT = 3;
-	public static final int PARTICLE_SPAWN_ATTEMPTS = 3;
 
 	public PaleMushroomBlock(ResourceKey<Feature> feature, Properties properties) {
 		super(feature, properties);
@@ -45,27 +40,7 @@ public class PaleMushroomBlock extends MushroomBlock {
 
 	@Override
 	public void animateTick(BlockState state, Level level, BlockPos pos, RandomSource random) {
-		if (!isActive(level, pos)) return;
-
-		final BlockPos.MutableBlockPos mutable = new BlockPos.MutableBlockPos();
-		for (int l = 0; l < PARTICLE_SPAWN_ATTEMPTS; ++l) {
-			if (random.nextFloat() >= 0.25F) continue;
-			mutable.setWithOffset(
-				pos,
-				Mth.nextInt(random, MIN_PARTICLE_SPAWN_WIDTH, MAX_PARTICLE_SPAWN_WIDTH),
-				Mth.nextInt(random, MIN_PARTICLE_SPAWN_HEIGHT, MAX_PARTICLE_SPAWN_HEIGHT),
-				Mth.nextInt(random, MIN_PARTICLE_SPAWN_WIDTH, MAX_PARTICLE_SPAWN_WIDTH)
-			);
-			final BlockState insideState = level.getBlockState(mutable);
-			if (insideState.isCollisionShapeFullBlock(level, mutable) || level.isRainingAt(mutable)) continue;
-			level.addParticle(
-				WWParticleTypes.PALE_SPORE.get(),
-				mutable.getX() + random.nextDouble(),
-				mutable.getY() + random.nextDouble(),
-				mutable.getZ() + random.nextDouble(),
-				0D, 0D, 0D
-			);
-		}
+		if (!ModLoader.isClient() || !isActive(level, pos)) return;
+		PaleParticleSpawners.PALE_SPORE_SPAWNER.tick(level, pos, random);
 	}
-
 }